raised warning level to 4 on own projects, lowered it to 1 on 3rdparty and external code
git-svn-id: https://bugengine.googlecode.com/svn/trunk@525 f3700e90-1ba7-11de-b56b-c9554f25afe1

+++ src/game/discworld/src/main.cc
@@ -25,7 +25,7 @@
 
 static BugEngine::Value buildRenderTarget()
 {
-    ref<BugEngine::Graphics::RenderWindow> result = ref<BugEngine::Graphics::RenderWindow>::create(BugEngine::gameArena(), 800, 600, "discworld v0.1", false);
+    ref<BugEngine::Graphics::RenderWindow> result = ref<BugEngine::Graphics::RenderWindow>::create(BugEngine::gameArena(), (u16)800, (u16)600, "discworld v0.1", false);
     return BugEngine::Value(result);
 }
